Updated: Moved the exceptions to a package in the domain of the controllers.
