+++ backend/src/main/java/br/balladesh/pjcappbackend/controllers/api/albums/GetFilterAlbumsController.java
@@ -4,8 +4,8 @@ import br.balladesh.pjcappbackend.dto.api.albums.PagedAlbumResponseBody;
 import br.balladesh.pjcappbackend.entity.AlbumEntity;
 import br.balladesh.pjcappbackend.repository.AlbumRepository;
 import br.balladesh.pjcappbackend.utilities.factories.CreateResponseFromExceptionFactory;
-import br.balladesh.pjcappbackend.utilities.errors.BadRequestException;
-import br.balladesh.pjcappbackend.utilities.errors.InternalServerErrorException;
+import br.balladesh.pjcappbackend.controllers.exceptions.BadRequestException;
+import br.balladesh.pjcappbackend.controllers.exceptions.InternalServerErrorException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
