+++ backend/src/main/java/br/balladesh/pjcappbackend/controllers/exceptions/NotFoundException.java
@@ -1,8 +1,9 @@
-package br.balladesh.pjcappbackend.utilities.errors;
+package br.balladesh.pjcappbackend.controllers.exceptions;
 
+import br.balladesh.pjcappbackend.utilities.errors.HttpException;
 import org.springframework.http.HttpStatus;
 
-public class NotFoundException extends HttpException{
+public class NotFoundException extends HttpException {
   public NotFoundException() {
     super(HttpStatus.NOT_FOUND);
   }
